Исправил классы базы данных dsu. Добавил получение данных о факультетах, направлениях, группах и студентах

+++ DSUContextDBService/Models/CaseCEdue.cs
@@ -2,7 +2,7 @@
 {
     public partial class CaseCEdue
     {
-        public short EduesId { get; set; }
+        public int EduesId { get; set; }
         public string? EduesName { get; set; }
     }
 }
