+++ DSUContextDBService/Models/CaseCEdukind.cs
@@ -2,9 +2,9 @@
 {
     public partial class CaseCEdukind
     {
-        public short EdukindId { get; set; }
+        public int EdukindId { get; set; }
         public string? Edukind { get; set; }
-        public short? Yearedu { get; set; }
+        public int? Yearedu { get; set; }
         public string? Abr { get; set; }
     }
 }
